Include duckdb distance in dev releases
Right now nightlies only get published to PyPi when there are actual
python code changes, because the publishing script won't publish
packages with a version string that already exists in PyPi. Sometimes
changes only happen on the duckdb repo though, and not on duckdb-python.
This change includes the number of duckdb changes in the version number
too. That way whenever one of the two gets updated a new python package
is published.

Similar to #170