Add windows_arm64 wheel build
This change adds Windows ARM64 platform for wheel builds.

The build is confirmed to work, but I am unable to make the tests
pass on ARM64 - included changes in `pyproject.toml` are likely
incomplete.

The tests are failing for me on AMD64 too, so it is possible that the
problem is not specific to ARM64, thus submitting the PR in this
incomplete state. Would appreciate pointers how to improve it.

Ref: #duckdblabs/duckdb-internal#5891